Refactor: 확장성 고려한 코드 변경 및 디렉토리 구조 변경에 따른 코드 수정
